port the old cli over to the new opup crate
