+++ bin/opup/Cargo.toml
@@ -18,6 +18,11 @@ ethabi = "18.0.0"
 num-bigint = "0.4"
 thiserror = "1.0"
 
+eyre = "0.6.8"
+inquire = "0.6.2"
+serde = "1.0.163"
+serde_json = "1.0.96"
+
 futures.workspace = true
 
 clap = { version = "4.4.3", features = ["derive"] }
