+++ bin/opup/src/banners.rs
@@ -1,0 +1,19 @@
+use eyre::Result;
+
+/// Print an op-up banner.
+pub fn banner() -> Result<()> {
+    println!(
+        r#"                        
+                                         
+        ___   _____            __  __  _____   
+       / __`\/\ '__`\  _______/\ \/\ \/\ '__`\ 
+      /\ \L\ \ \ \L\ \/\______\ \ \_\ \ \ \L\ \
+      \ \____/\ \ ,__/\/______/\ \____/\ \ ,__/
+       \/___/  \ \ \/           \/___/  \ \ \/ 
+                \ \_\                    \ \_\ 
+                 \/_/                     \/_/ 
+      
+      "#
+    );
+    Ok(())
+}
