+++ bin/opup/src/etc/clock.rs
@@ -1,0 +1,9 @@
+use std::time::{Duration, SystemTime, UNIX_EPOCH};
+
+/// Returns the current timestamp in seconds.
+pub fn current_timestamp() -> u64 {
+    SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .unwrap()
+        .as_secs()
+}
