+++ bin/opup/src/etc/mod.rs
@@ -3,3 +3,21 @@ pub mod runner;
 
 /// Telemetry contains helpers for initializing telemetry.
 pub mod telemetry;
+
+/// Contains helper functions for shell commands.
+pub mod commands;
+
+/// Contains helper functions for git commands.
+pub mod git;
+
+/// Contains helper functions for networking.
+pub mod net;
+
+/// Contains JSON helpers.
+pub mod json;
+
+/// System clock helpers.
+pub mod clock;
+
+/// Various macros.
+pub mod macros;
