+++ bin/opup/src/lib.rs
@@ -6,3 +6,12 @@ pub mod stack;
 
 /// General purpose helpers.
 pub mod etc;
+
+/// System-level constants.
+pub mod constants;
+
+/// Deployment addresses.
+pub mod addresses;
+
+/// Command banners.
+pub mod banners;
